Library

Velocty Structure Function Tools

TurbulentAnalysis.umoments — Function

umoments(ur, Δr, x, rmax; min_points = 4, calculate_sf3 = true)

Calculate the second moments of longitudinal velocity ur with separation Δr and coordinate x using all possible separations in the region rmax around x.

min_points is the minimum number of values of ur to to use to calculate the velocity moments (default is 4, resulting in 3 separations). Optionally calculate the third moments of ur by setting calculate_sf3 = true (default).

TurbulentAnalysis.ϵofx — Function

ϵofx(umom; fittype=:OLS, γ = 0.32, minp2fit = 3, exponent_correction = 0, calculate_sf3 = true)

Calculate TKE dissipation rate ϵ from the velocity moments returned by umoments, and the doppler noise variance.

Calculate from both the second and third order moments.

...

Arguments

fittype: The method used for fitting the line (:OLS or :GLS).

γ: The confidence level (i.e. γ-th percentile).

minp2fit: The minimum number of points to fit to.

exponent_correction: optional correction factor for the exponent.

calculate_sf3: Whether to calculate ϵ from the third-order moments (default is true).

...

TurbulentAnalysis.ϵofxandt — Function

ϵofxandt(ur, x, t, Δr, fs, Δt, rmax; xmin = minimum(x), xmax = maximum(x), min_points = 4, fittype = :OLS, γ = 0.32, minp2fit = 3, calculate_sf3 = true)

Calculate a TKE dissipation rate time series using velocities ur with separation Δr, spatial coordinate x, times t, and downsample time resolution Δt.

...

Arguments

ur: The matrix of velocities.

x: The spatial coordinates of ur.

t: The times of ur.

Δr: The separation distance between velocities.

fs: the sample frequency of t.

Δt: the time interval to average velocity moments over. This becomes the resulting time resolution.

rmax: the maximum range separation between velocities to consider.

xmin: The minimum value in x to include in the calculations. Can either be a single value or a vector of size(t).

xmax: The maximum value in x to include in the calculations. Can either be a single value or a vector of size(t).

min_points: The minimum number of values of ur to use to calculate ϵ (default is 4, resulting in 3 separations).

fittype: The method used for fitting the line (:OLS or :GLS).

γ: The confidence level (i.e. γ-th percentile).

minp2fit: The minimum number of points to fit to.

exponent_correction: optional correction factor for the exponent.

calculate_sf3: Whether to calculate ϵ from the third-order moments (default is true).

...

Utils

TurbulentAnalysis.DLLx — Function

DLLx(r; exponent_correction = 0)

Change of variables of ranges r to x values for a linear fit for DLL. Optionally specify a exponent_correction to account for different PDF of the breakage coefficient.

TurbulentAnalysis.DLLLx — Function

DLLLx(r; exponent_correction = 0)

Change of variables of ranges r to x values for a linear fit for DLLL. Optionally specify a exponent_correction to account for different PDF of the breakage coefficient.

TurbulentAnalysis.ϵ_DLL — Function

ϵ_DLL(A; C₂ = 2.0)

Calculate dissipation rate ϵ from the slope A of a linear fit to DLL.

TurbulentAnalysis.ϵ_DLLL — Function

ϵ_DLLL(A; C₂ = 2.0)

Calculate dissipation rate ϵ from the slope A of a linear fit to DLLL.

TurbulentAnalysis.σ_DLL — Function

σ_DLL(B)

Calculate the noise variance from the intercept of the fit to DLL.

TurbulentAnalysis.TimeDuration — Function

TimeDuration(t)

Calculate the duration of the time series t.

TimeDuration(t)

Calculate the duration of the time series t in seconds.

TurbulentAnalysis.xlims2idx — Function

xlims2idx(x, xmin, xmax)

Create a vector of ranges from xmin to xmax in x. xlims2idx x can either be a vector or a matrix, while xmin and xmax can be a single value or a vector with size size(x)[2]. The vector must be 1xtime so the dimensions line up.

Quality Control Tools

TurbulentAnalysis.FailsQC — Function

FailsQC(q, n)

Determine if a point with QC flag q fails quality control condition number n.

TurbulentAnalysis.QCϵts — Function

QCϵts(ϵ_ts, max_relative_uncertainty, max_relative_power_diff; exponent_correction = 0)

Replace values of ϵ in ϵ_ts that exceed the max_relative_uncertainty or max_relative_power_diff with fillvalue. Optionally specify a exponent_correction to account for different PDF of the breakage coefficient.

Misc

TurbulentAnalysis.largest_eddies — Function

largest_eddies(u, x, t, Δr, δt, fs; xmin = minimum(x), xmax = maximum(x), x0_c = nothing, SF = nothing, logmean=false)

Estimate the properties of the largest eddies of u.

...

Arguments

u: longitudinal velocities along the line defined by x. t is the second dimension.

x: The spatial coordinates of u.

t: The times of u.

Δr: The separation distance between velocities.

δt: the time interval to average over.

fs: the sample frequency of t (units of 1/δt).

xmin: The minimum value in x to include in the calculations. Can either be a single value or a vector of size(t).

xmax: The maximum value in x to include in the calculations. Can either be a single value or a vector of size(t).

x0_c: optional reference position to calculate covariance from.

SF: optional structure function ϵ object to average to compare.

logmean: Whether to take the logarithmic mean of the SF ϵ estimates.

...